refactor the scrapper file and fix a bug
